--- tests/codes/eurocode/en_1993_1_1_2025/__init__.py
@@ -1,1 +1,0 @@
-"""Package contains tests for Eurocode EN 1993-1-1:2025."""
